migrating some IT to BaseSepInProcessTest
